Fix compil

+++ src/Transformable/Camera/Camera.cpp
@@ -24,8 +24,8 @@ namespace Transformable {
             _width = width;
             _height = height;
         }
-        std::vector<Point3f> Camera::computeAxis() {
-            std::vector<Point3f> axis;
+        std::vector<Point3d> Camera::computeAxis() {
+            std::vector<Point3d> axis;
             //for the V1 of the camera, we create a rectangle
             //the vectors do not start from the same origin, they go straight to the camera
